+++ docs/changelog.md
@@ -2,6 +2,7 @@
 ======================================================
 **API 2.0.0-SNAPSHOT**
 **Requires Journeymap v6.0.0**
+* Added Server Properties, so mods can add properties to JourneyMap's Server Admin options tab.
 * Added LayeredScreen abstraction for layirng screens.
 * Refactored Waypoint and WaypointGroups to be more concrete. 
 * Refactored events to be loader agnostic.